fixes abstraction
